Updated paths for creating release version, updated testnet scripts, added quick quide for configuring nodes on testnet

+++ scripts/testnet.sh
@@ -3,6 +3,16 @@ set -e
 
 CHAIN_ID="symphony-testnet-1"
 
+if [ "$#" -ne 4 ]; then
+    echo "Usage: $0 <NODE1_API> <NODE2_API> <NODE3_API> <NODE4_API>"
+    exit 1
+fi
+
+NODE1_API=$1
+NODE2_API=$2
+NODE3_API=$3
+NODE4_API=$4
+
 killall symphonyd || true
 rm -rf $HOME/.symphonyd/
 
@@ -85,24 +95,13 @@ symphonyd add-genesis-account $(symphonyd keys show validator1 -a --keyring-back
 symphonyd gentx validator1 5000000000note --moniker="validator1" --chain-id= --commission-rate="0.1" --commission-max-rate="0.2" --commission-max-change-rate="0.05" --min-self-delegation="500000000" --keyring-backend=test --home=$HOME/.symphonyd/validator1 --chain-id=$CHAIN_ID
 symphonyd collect-gentxs --home=$HOME/.symphonyd/validator1
 
-# # copy validator1 genesis file to validator2-4
-# cp $HOME/.symphonyd/validator1/config/genesis.json $HOME/.symphonyd/validator2/config/genesis.json
-# cp $HOME/.symphonyd/validator1/config/genesis.json $HOME/.symphonyd/validator3/config/genesis.json
-# cp $HOME/.symphonyd/validator1/config/genesis.json $HOME/.symphonyd/validator4/config/genesis.json
+# copy validator1 genesis file to validator2-4
+ cp $HOME/.symphonyd/validator1/config/genesis.json $HOME/.symphonyd/validator2/config/genesis.json
+ cp $HOME/.symphonyd/validator1/config/genesis.json $HOME/.symphonyd/validator3/config/genesis.json
+ cp $HOME/.symphonyd/validator1/config/genesis.json $HOME/.symphonyd/validator4/config/genesis.json
 
-# symphonyd add-genesis-account $(symphonyd keys show validator3 -a --keyring-backend=test --home=$HOME/.symphonyd/validator3) 100000000000note --home=$HOME/.symphonyd/validator3
-# symphonyd gentx validator3 500000000note --moniker="validator3" --chain-id=$CHAIN_ID --commission-rate="0.1" --commission-max-rate="0.2" --commission-max-change-rate="0.05" --min-self-delegation="500000000" --keyring-backend=test --home=$HOME/.symphonyd/validator3 --chain-id=$CHAIN_ID
-# symphonyd collect-gentxs --home=$HOME/.symphonyd/validator3
-
-# symphonyd add-genesis-account $(symphonyd keys show validator4 -a --keyring-backend=test --home=$HOME/.symphonyd/validator4) 100000000000note --home=$HOME/.symphonyd/validator4
-# symphonyd gentx validator4 500000000note --moniker="validator4" --chain-id=$CHAIN_ID --commission-rate="0.1" --commission-max-rate="0.2" --commission-max-change-rate="0.05" --min-self-delegation="500000000" --keyring-backend=test --home=$HOME/.symphonyd/validator4 --chain-id=$CHAIN_ID
-# symphonyd collect-gentxs --home=$HOME/.symphonyd/validator4
-
-# port key (validator1 uses default ports)
-# validator1 1317, 9050, 9091, 26658, 26657, 26656, 6060, 26660
-# validator2 1316, 9088, 9089, 26655, 26654, 26653, 6061, 26630
-# validator3 1315, 9086, 9087, 26652, 26651, 26650, 6062, 26620
-# validator4 1314, 9084, 9085, 26649, 26648, 26647, 6063, 26610
+# port key (validators uses default ports)
+# validator 1317, 9050, 9091, 26658, 26657, 26656, 6060, 26660
 
 # change app.toml values
 VALIDATOR1_APP_TOML=$HOME/.symphonyd/validator1/config/app.toml
@@ -110,24 +109,13 @@ VALIDATOR2_APP_TOML=$HOME/.symphonyd/validator2/config/app.toml
 VALIDATOR3_APP_TOML=$HOME/.symphonyd/validator3/config/app.toml
 VALIDATOR4_APP_TOML=$HOME/.symphonyd/validator4/config/app.toml
 
-# validator1
-sed -i -E 's|0.0.0.0:9090|0.0.0.0:9050|g' $VALIDATOR1_APP_TOML
-
 # validator2
-sed -i -E 's|localhost:1317|localhost:1316|g' $VALIDATOR2_APP_TOML
-sed -i -E 's|localhost:9090|localhost:9088|g' $VALIDATOR2_APP_TOML
-sed -i -E 's|localhost:9091|localhost:9089|g' $VALIDATOR2_APP_TOML
+sed -i -E 's|adaptive-fee-enabled = "false"|adaptive-fee-enabled = "true"|g' $VALIDATOR3_APP_TOML
 
 # validator3
-sed -i -E 's|localhost:1317|localhost:1315|g' $VALIDATOR3_APP_TOML
-sed -i -E 's|localhost:9090|localhost:9086|g' $VALIDATOR3_APP_TOML
-sed -i -E 's|localhost:9091|localhost:9087|g' $VALIDATOR3_APP_TOML
 sed -i -E 's|adaptive-fee-enabled = "false"|adaptive-fee-enabled = "true"|g' $VALIDATOR3_APP_TOML
 
 # validator4
-sed -i -E 's|localhost:1317|localhost:1314|g' $VALIDATOR4_APP_TOML
-sed -i -E 's|localhost:9090|localhost:9084|g' $VALIDATOR4_APP_TOML
-sed -i -E 's|localhost:9091|localhost:9085|g' $VALIDATOR4_APP_TOML
 sed -i -E 's|adaptive-fee-enabled = "false"|adaptive-fee-enabled = "true"|g' $VALIDATOR4_APP_TOML
 
 
@@ -139,74 +127,40 @@ VALIDATOR4_CONFIG=$HOME/.symphonyd/validator4/config/config.toml
 
 # validator1
 sed -i -E 's|allow_duplicate_ip = false|allow_duplicate_ip = true|g' $VALIDATOR1_CONFIG
-# sed -i -E 's|version = "v0"|version = "v1"|g' $VALIDATOR1_CONFIG
 sed -i -E 's|prometheus = false|prometheus = true|g' $VALIDATOR1_CONFIG
 
 # validator2
-sed -i -E 's|tcp://127.0.0.1:26658|tcp://127.0.0.1:26655|g' $VALIDATOR2_CONFIG
-sed -i -E 's|tcp://127.0.0.1:26657|tcp://127.0.0.1:26654|g' $VALIDATOR2_CONFIG
-sed -i -E 's|tcp://0.0.0.0:26656|tcp://0.0.0.0:26653|g' $VALIDATOR2_CONFIG
 sed -i -E 's|allow_duplicate_ip = false|allow_duplicate_ip = true|g' $VALIDATOR2_CONFIG
 sed -i -E 's|prometheus = false|prometheus = true|g' $VALIDATOR2_CONFIG
-sed -i -E 's|prometheus_listen_addr = ":26660"|prometheus_listen_addr = ":26630"|g' $VALIDATOR2_CONFIG
 
 # validator3
-sed -i -E 's|tcp://127.0.0.1:26658|tcp://127.0.0.1:26652|g' $VALIDATOR3_CONFIG
-sed -i -E 's|tcp://127.0.0.1:26657|tcp://127.0.0.1:26651|g' $VALIDATOR3_CONFIG
-sed -i -E 's|tcp://0.0.0.0:26656|tcp://0.0.0.0:26650|g' $VALIDATOR3_CONFIG
 sed -i -E 's|allow_duplicate_ip = false|allow_duplicate_ip = true|g' $VALIDATOR3_CONFIG
 sed -i -E 's|prometheus = false|prometheus = true|g' $VALIDATOR3_CONFIG
-sed -i -E 's|prometheus_listen_addr = ":26660"|prometheus_listen_addr = ":26620"|g' $VALIDATOR3_CONFIG
 
 # validator4
-sed -i -E 's|tcp://127.0.0.1:26658|tcp://127.0.0.1:26649|g' $VALIDATOR4_CONFIG
-sed -i -E 's|tcp://127.0.0.1:26657|tcp://127.0.0.1:26648|g' $VALIDATOR4_CONFIG
-sed -i -E 's|tcp://0.0.0.0:26656|tcp://0.0.0.0:26647|g' $VALIDATOR4_CONFIG
 sed -i -E 's|allow_duplicate_ip = false|allow_duplicate_ip = true|g' $VALIDATOR4_CONFIG
 sed -i -E 's|prometheus = false|prometheus = true|g' $VALIDATOR4_CONFIG
-sed -i -E 's|prometheus_listen_addr = ":26660"|prometheus_listen_addr = ":26610"|g' $VALIDATOR4_CONFIG
-
 
 # persistent_peers
 
 # copy tendermint node id of validator 2,3,4 to 1
-sed -i -E "s|persistent_peers = \"\"|persistent_peers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@localhost:26654,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@localhost:26651,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@localhost:26648\"|g" $HOME/.symphonyd/validator1/config/config.toml
+sed -i -E "s|persistent_peers = \"\"|persistent_peers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@$NODE2_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@$NODE3_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@$NODE4_API:26656\"|g" $HOME/.symphonyd/validator1/config/config.toml
 # copy tendermint node id of 1,3,4 to 2
-sed -i -E "s|persistent_peers = \"\"|persistent_peers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@localhost:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@localhost:26651,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@localhost:26648\"|g" $HOME/.symphonyd/validator2/config/config.toml
+sed -i -E "s|persistent_peers = \"\"|persistent_peers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@$NODE1_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@$NODE3_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@$NODE4_API:26656\"|g" $HOME/.symphonyd/validator2/config/config.toml
 # copy tendermint node id of 1,2,4 to 3
-sed -i -E "s|persistent_peers = \"\"|persistent_peers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@localhost:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@localhost:26654,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@localhost:26648\"|g" $HOME/.symphonyd/validator3/config/config.toml
+sed -i -E "s|persistent_peers = \"\"|persistent_peers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@$NODE1_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@$NODE2_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@$NODE4_API:26656\"|g" $HOME/.symphonyd/validator3/config/config.toml
 # copy tendermint node id of 1,2,3 to 4
-sed -i -E "s|persistent_peers = \"\"|persistent_peers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@localhost:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@localhost:26654,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@localhost:26651\"|g" $HOME/.symphonyd/validator4/config/config.toml
+sed -i -E "s|persistent_peers = \"\"|persistent_peers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@$NODE1_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@$NODE2_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@$NODE3_API:26656\"|g" $HOME/.symphonyd/validator4/config/config.toml
 
 # rpc_servers
 
 # copy tendermint node id  2,3,4 to 1
-sed -i -E "s|rpc_servers = \"\"|rpc_servers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@localhost:26654,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@localhost:26651,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@localhost:26648\"|g" $HOME/.symphonyd/validator1/config/config.toml
+sed -i -E "s|rpc_servers = \"\"|rpc_servers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@$NODE2_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@$NODE3_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@$NODE4_API:26656\"|g" $HOME/.symphonyd/validator1/config/config.toml
 # copy tendermint node id of 1,3,4 to 2
-sed -i -E "s|rpc_servers = \"\"|rpc_servers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@localhost:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@localhost:26651,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@localhost:26648\"|g" $HOME/.symphonyd/validator2/config/config.toml
+sed -i -E "s|rpc_servers = \"\"|rpc_servers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@$NODE1_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@$NODE3_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@$NODE4_API:26656\"|g" $HOME/.symphonyd/validator2/config/config.toml
 # copy tendermint node id of 1,2,4 to 3
-sed -i -E "s|rpc_servers = \"\"|rpc_servers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@localhost:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@localhost:26654,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@localhost:26648\"|g" $HOME/.symphonyd/validator3/config/config.toml
+sed -i -E "s|rpc_servers = \"\"|rpc_servers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@$NODE1_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@$NODE2_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator4)@$NODE4_API:26656\"|g" $HOME/.symphonyd/validator3/config/config.toml
 # copy tendermint node id of 1,2,3 to 4
-sed -i -E "s|rpc_servers = \"\"|rpc_servers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@localhost:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@localhost:26654,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@localhost:26651\"|g" $HOME/.symphonyd/validator4/config/config.toml
+sed -i -E "s|rpc_servers = \"\"|rpc_servers = \"$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator1)@$NODE1_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator2)@$NODE2_API:26656,$(symphonyd tendermint show-node-id --home=$HOME/.symphonyd/validator3)@$NODE3_API:26656\"|g" $HOME/.symphonyd/validator4/config/config.toml
 
-echo "All 4 Validators are up!"
-
-# copy validator1 genesis file to validator2-4
-
-# start separately all 4 validators on different nodes
-# symphonyd start --home=$HOME/.symphonyd
-
-# send note to other validators
-# send note from 1-st validator to 2-nd validator
-# symphonyd tx bank send validator1 $(symphonyd keys show validator2 -a --keyring-backend=test --home=$HOME/.symphonyd/validator2) 500000000note --keyring-backend=test --home=$HOME/.symphonyd/validator1 --chain-id=$CHAIN_ID --broadcast-mode sync --yes --fees 1000000note
-
-# send note from 1-st validator to 3-rd validator
-# symphonyd tx bank send validator1 $(symphonyd keys show validator3 -a --keyring-backend=test --home=$HOME/.symphonyd/validator3) 400000000note --keyring-backend=test --home=$HOME/.symphonyd/validator1 --chain-id=$CHAIN_ID --broadcast-mode sync --yes --fees 1000000note
-
-# send note from 1-st validator to 4-rd validator
-# symphonyd tx bank send validator1 $(symphonyd keys show validator4 -a --keyring-backend=test --home=$HOME/.symphonyd/validator4) 400000000note --keyring-backend=test --home=$HOME/.symphonyd/validator1 --chain-id=$CHAIN_ID --broadcast-mode sync --yes --fees 1000000note
-
-# add 2- 4 validators
-# symphonyd tx staking create-validator --amount=500000000note --from=validator1 --pubkey=$(symphonyd tendermint show-validator --home=$HOME/.symphonyd/validator2) --moniker="validator2" --chain-id=$CHAIN_ID --commission-rate="0.1" --commission-max-rate="0.2" --commission-max-change-rate="0.05" --min-self-delegation="500000000" --keyring-backend=test --home=$HOME/.symphonyd/validator1 --broadcast-mode sync  --yes --fees 1000000note
-# symphonyd tx staking create-validator --amount=500000000note --from=validator1 --pubkey=$(symphonyd tendermint show-validator --home=$HOME/.symphonyd/validator3) --moniker="validator3" --chain-id=$CHAIN_ID --commission-rate="0.1" --commission-max-rate="0.2" --commission-max-change-rate="0.05" --min-self-delegation="500000000" --keyring-backend=test --home=$HOME/.symphonyd/validator1 --broadcast-mode sync  --yes --fees 1000000note
-# symphonyd tx staking create-validator --amount=500000000note --from=validator1 --pubkey=$(symphonyd tendermint show-validator --home=$HOME/.symphonyd/validator4) --moniker="validator4" --chain-id=$CHAIN_ID --commission-rate="0.1" --commission-max-rate="0.2" --commission-max-change-rate="0.05" --min-self-delegation="500000000" --keyring-backend=test --home=$HOME/.symphonyd/validator1 --broadcast-mode sync  --yes --fees 1000000note
+echo "All 4 Validators configuration are created!"
